# src/main/resources/com/sonicle/webtop/mail/openapi-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "v1",
    "title": "WebTop Mail",
    "description": "## Request header: X-WT-Mail-Machinery\nAny `/me/*` request may include the header `X-WT-Mail-Machinery: full` to declare that the caller (e.g. a mobile mail app) needs the user's full mail machinery (accounts, folder caches, IMAP idle) running server-side. With it, folder listings and unread counts are served from the live folder cache (the same values the WebTop web client shows) and message listings ride the warm cached paths. The first call bearing the header may take a few seconds while the machinery starts; subsequent calls are fast. The machinery is torn down automatically after ~15 minutes without any call for that user; the next call bearing the header transparently restarts it, so clients should simply send the header on every request and need no eviction handling. Callers that do not need live data (bulk integrations, occasional automation) should omit the header: they are served through per-call paths and never spin up per-user machinery.\n\n---\nGenerator notes:\nhttps://github.com/OpenAPITools/openapi-generator/issues/10010\nhttps://github.com/stoplightio/prism/discussions/523\nhttps://github.com/OpenAPITools/openapi-generator/issues/3172",
    "contact": {
      "email": "dev-team@sonicle.com"
    }
  },
...
  "security": [
    {
      "basicAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/settings/external-archiving": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get configuration data",
        "operationId": "getExternalArchivingConfiguration",
        "parameters": [
          {
            "name": "targetProfileId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalArchivingSettings"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "ExternalArchivingSettings": {
        "type": "object",
        "required": [
          "enabled",
          "host",
          "port",
          "protocol",
          "type"
        ],
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "Enabled status"
          },
          "type": {
            "type": "string",
            "description": "Archiving type"
          },
          "host": {
            "type": "string",
            "description": "IMAP server host"
          },
          "port": {
            "type": "integer",
            "description": "IMAP server port"
          },
          "protocol": {
            "type": "string",
            "description": "IMAP server protocol"
          },
          "username": {
            "type": "string",
            "description": "Account username"
          },
          "password": {
            "type": "string",
            "description": "Account password"
          },
          "folderPrefix": {
            "type": "string",
            "description": "IMAP folder prefix"
          },
          "minage": {
            "type": "integer",
            "description": "Archive messages older than days"
          }
        }
      },
...
    },
...
  }
}